requirements.txt
absl-py==2.3.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
google-api-core==2.28.1
google-auth==2.47.0
google-cloud-speech==2.35.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
idna==3.11
jax==0.7.1
jaxlib==0.7.1
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.21
ml_dtypes==0.5.4
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==12.1.0
pipdeptree==2.30.0
proto-plus==1.27.0
protobuf==4.25.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyaudio==0.2.11
pycparser==2.23
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
scipy==1.16.3
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.3
typing_extensions==4.15.0
urllib3==2.6.2
